arm/configs: bullhead: remove SysV IPC from kernel
System V IPCs are not compliant with Android's application lifecycle
because allocated resources are not freeable by the low memory killer.
This lead to global kernel resource leakage.

For example, there is no way to automatically release a SysV
semaphore allocated in the kernel when:
- a buggy or malicious process exits
- a non-buggy and non-malicious process crashes or is explicitly
  killed.

Killing processes automatically to make room for new ones is an
important part of Android's application lifecycle implementation.
This means that, even assuming only non-buggy and non-malicious
code, it is very likely that over time, the kernel global tables
used to implement SysV IPCs will fill up.

Bug: 24551430
Bug: 22300191

Signed-off-by: Patrick Tjin <pattjin@google.com>
Change-Id: I3ee30e0c3526114771479338817c725a1933fa59

Signed-off-by: Akhil Narang <something@akhilnarang.xyz>

Conflicts:
	arch/arm64/configs/bullhead_defconfig